Trait foreign_types_shared::ForeignTypeRef

pub trait ForeignTypeRef: Sized {
    type CType;
    unsafe fn from_ptr<'a>(ptr: *mut Self::CType) -> &'a Self { ... }

    unsafe fn from_ptr_mut<'a>(ptr: *mut Self::CType) -> &'a mut Self { ... }

    fn as_ptr(&self) -> *mut Self::CType { ... }
}

A trait implemented by types which reference borrowed foreign types.

Associated Types

type CType

The raw C type.

Loading content...
Provided methods

unsafe fn from_ptr<'a>(ptr: *mut Self::CType) -> &'a Self

Constructs a shared instance of this type from its raw type.

unsafe fn from_ptr_mut<'a>(ptr: *mut Self::CType) -> &'a mut Self

Constructs a mutable reference of this type from its raw type.

fn as_ptr(&self) -> *mut Self::CType

Returns a raw pointer to the wrapped value.

Loading content...